requirements.txt
asgiref==3.7.2
colorama==0.4.6
Django==4.2.2
Pillow==10.0.0
pypng==0.20220715.0
qrcode==7.4.2
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
